+++ packages/dify-ui/src/radio/index.tsx
@@ -1,23 +1,29 @@
 'use client'
 
 import type { Radio as BaseRadioNS } from '@base-ui/react/radio'
+import type { RadioGroup as BaseRadioGroupNS } from '@base-ui/react/radio-group'
 import type * as React from 'react'
 import { Radio as BaseRadio } from '@base-ui/react/radio'
+import { RadioGroup as BaseRadioGroup } from '@base-ui/react/radio-group'
 import { cn } from '../cn'
 
-const radioControlClassName = cn(
-  'inline-flex size-4 shrink-0 touch-manipulation items-center justify-center rounded-full p-0 transition-colors motion-reduce:transition-none',
-  'border border-components-radio-border bg-components-radio-bg shadow-xs shadow-shadow-shadow-3',
-  'hover:border-components-radio-border-hover hover:bg-components-radio-bg-hover',
-  'focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-state-accent-solid focus-visible:ring-offset-0',
-  'data-checked:border-[5px] data-checked:border-components-radio-border-checked data-checked:hover:border-components-radio-border-checked-hover',
-  'data-disabled:cursor-not-allowed data-disabled:border-components-radio-border-disabled data-disabled:bg-components-radio-bg-disabled',
-  'data-disabled:hover:border-components-radio-border-disabled data-disabled:hover:bg-components-radio-bg-disabled',
-  'data-disabled:data-checked:border-[5px] data-disabled:data-checked:border-components-radio-border-checked-disabled',
-  'data-disabled:data-checked:hover:border-components-radio-border-checked-disabled',
-)
+export type RadioGroupProps<Value = string>
+  = Omit<BaseRadioGroupNS.Props<Value>, 'className'>
+    & {
+      className?: string
+    }
 
-const radioSkeletonClassName = 'size-4 shrink-0 rounded-full bg-text-quaternary opacity-20'
+export function RadioGroup<Value = string>({
+  className,
+  ...props
+}: RadioGroupProps<Value>) {
+  return (
+    <BaseRadioGroup<Value>
+      className={cn('flex items-center gap-2', className)}
+      {...props}
+    />
+  )
+}
 
 export type RadioItemProps<Value = string>
   = Omit<BaseRadioNS.Root.Props<Value>, 'className'>
@@ -51,7 +57,18 @@ export function RadioControl({
     <BaseRadio.Indicator
       {...props}
       keepMounted
-      className={cn(radioControlClassName, className)}
+      className={cn(
+        'inline-flex size-4 shrink-0 touch-manipulation items-center justify-center rounded-full p-0 transition-colors motion-reduce:transition-none',
+        'border border-components-radio-border bg-components-radio-bg shadow-xs shadow-shadow-shadow-3',
+        'hover:border-components-radio-border-hover hover:bg-components-radio-bg-hover',
+        'focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-state-accent-solid focus-visible:ring-offset-0',
+        'data-checked:border-[5px] data-checked:border-components-radio-border-checked data-checked:hover:border-components-radio-border-checked-hover',
+        'data-disabled:cursor-not-allowed data-disabled:border-components-radio-border-disabled data-disabled:bg-components-radio-bg-disabled',
+        'data-disabled:hover:border-components-radio-border-disabled data-disabled:hover:bg-components-radio-bg-disabled',
+        'data-disabled:data-checked:border-[5px] data-disabled:data-checked:border-components-radio-border-checked-disabled',
+        'data-disabled:data-checked:hover:border-components-radio-border-checked-disabled',
+        className,
+      )}
     />
   )
 }
@@ -65,7 +82,18 @@ export function Radio<Value = string>({
 }: RadioProps<Value>) {
   return (
     <BaseRadio.Root<Value>
-      className={cn(radioControlClassName, className)}
+      className={cn(
+        'inline-flex size-4 shrink-0 touch-manipulation items-center justify-center rounded-full p-0 transition-colors motion-reduce:transition-none',
+        'border border-components-radio-border bg-components-radio-bg shadow-xs shadow-shadow-shadow-3',
+        'hover:border-components-radio-border-hover hover:bg-components-radio-bg-hover',
+        'focus-visible:outline-hidden focus-visible:ring-2 focus-visible:ring-state-accent-solid focus-visible:ring-offset-0',
+        'data-checked:border-[5px] data-checked:border-components-radio-border-checked data-checked:hover:border-components-radio-border-checked-hover',
+        'data-disabled:cursor-not-allowed data-disabled:border-components-radio-border-disabled data-disabled:bg-components-radio-bg-disabled',
+        'data-disabled:hover:border-components-radio-border-disabled data-disabled:hover:bg-components-radio-bg-disabled',
+        'data-disabled:data-checked:border-[5px] data-disabled:data-checked:border-components-radio-border-checked-disabled',
+        'data-disabled:data-checked:hover:border-components-radio-border-checked-disabled',
+        className,
+      )}
       {...props}
     />
   )
@@ -79,7 +107,7 @@ export function RadioSkeleton({
 }: RadioSkeletonProps) {
   return (
     <div
-      className={cn(radioSkeletonClassName, className)}
+      className={cn('size-4 shrink-0 rounded-full bg-text-quaternary opacity-20', className)}
       {...props}
     />
   )
